processing/pyproject.toml
[project]
name = "processing"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.39",
    "pyspark>=4.1.1",
    "psycopg2-binary>=2.9.9",
    "pyarrow>=15.0.0",
    "python-dotenv>=1.0.0",
    "pandas"
]

[project.scripts]
load-postgres = "loader.cli:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["loader"]
